refactor(task-014): consolidar múltiplas versões de TransactionsPage
Consolida 3 versões diferentes de TransactionsPage em uma única versão canônica (TransactionsPageInfiniteScroll), removendo código duplicado e confusão.

Mudanças:
- Movido tipo Transaction para TransactionsPageInfiniteScroll
- Atualizado imports em TransactionModal, TransactionTable, TransactionFilters, TransactionModalImproved e TransactionsPageIntegrated
- Removido TransactionsPage.tsx (versão mock obsoleta)
- Removido TransactionsPageIntegrated.tsx (versão antiga integrada)
- Mantido TransactionsPageInfiniteScroll como versão canônica

Benefícios:
- Código mais limpo e organizado (467 linhas removidas)
- Redução de confusão sobre qual versão usar
- Manutenção mais fácil
- Versão única e canônica
- Todos os 316 testes continuam passando
